refactor(nps): extrair NpsScoreSelector para components/nps/
